Separate the http server from the task manager
